Move TCCPP-specific components to their own subdirectory
