+++ .gitmodules
@@ -1,11 +1,11 @@
-[submodule "wheatley-private"]
-	path = src/wheatley-private
-	url = https://github.com/TCCPP/empty.git
-	# url = git@github.com:TCCPP/wheatley-private.git
+# [submodule "src/modules/tccpp/private"]
+# 	path = src/modules/tccpp/private
+# 	url = git@github.com:TCCPP/wheatley-private.git
 
 # [submodule "dyno-import"]
 # 	path = dyno-import
 # 	url = git@github.com:TCCPP/dyno-import.git
+
 [submodule "wiki"]
 	path = wiki
 	url = git@github.com:TCCPP/wiki-articles.git
