+++ src/modules/tccpp/components/anti-autoreact.ts
@@ -1,8 +1,8 @@
 import * as Discord from "discord.js";
 import { strict as assert } from "assert";
-import { departialize } from "../utils/discord.js";
-import { M } from "../utils/debugging-and-logging.js";
-import { BotComponent } from "../bot-component.js";
+import { departialize } from "../../../utils/discord.js";
+import { M } from "../../../utils/debugging-and-logging.js";
+import { BotComponent } from "../../../bot-component.js";
 
 export default class AntiAutoreact extends BotComponent {
     obnoxious_autoreact_ids = new Set([
