+++ src/modules/tccpp/components/anti-screenshot.ts
@@ -2,11 +2,11 @@ import { strict as assert } from "assert";
 
 import * as Discord from "discord.js";
 
-import { delay } from "../utils/misc.js";
-import { M } from "../utils/debugging-and-logging.js";
-import { colors } from "../common.js";
-import { BotComponent } from "../bot-component.js";
-import { Wheatley } from "../wheatley.js";
+import { delay } from "../../../utils/misc.js";
+import { M } from "../../../utils/debugging-and-logging.js";
+import { colors } from "../../../common.js";
+import { BotComponent } from "../../../bot-component.js";
+import { Wheatley } from "../../../wheatley.js";
 
 const DISMISS_TIME = 30 * 1000;
