+++ src/modules/tccpp/components/c-help-redirect.ts
@@ -1,13 +1,13 @@
 import * as Discord from "discord.js";
 import { strict as assert } from "assert";
-import { BotComponent } from "../bot-component.js";
-import { CommandSetBuilder } from "../command-abstractions/command-set-builder.js";
-import { Wheatley } from "../wheatley.js";
-import { EarlyReplyMode, TextBasedCommandBuilder } from "../command-abstractions/text-based-command-builder.js";
-import { TextBasedCommand } from "../command-abstractions/text-based-command.js";
-import { HOUR, DAY } from "../common.js";
-import { M } from "../utils/debugging-and-logging.js";
-import { SelfClearingSet } from "../utils/containers.js";
+import { BotComponent } from "../../../bot-component.js";
+import { CommandSetBuilder } from "../../../command-abstractions/command-set-builder.js";
+import { Wheatley } from "../../../wheatley.js";
+import { EarlyReplyMode, TextBasedCommandBuilder } from "../../../command-abstractions/text-based-command-builder.js";
+import { TextBasedCommand } from "../../../command-abstractions/text-based-command.js";
+import { HOUR, DAY } from "../../../common.js";
+import { M } from "../../../utils/debugging-and-logging.js";
+import { SelfClearingSet } from "../../../utils/containers.js";
 
 const code_block_start = "```";
