+++ src/modules/tccpp/components/core-guidelines.ts
@@ -2,16 +2,16 @@ import * as Discord from "discord.js";
 
 import * as fs from "fs";
 
-import { M } from "../utils/debugging-and-logging.js";
+import { M } from "../../../utils/debugging-and-logging.js";
 
-import { Index, IndexEntry } from "../algorithm/search.js";
-import { core_guidelines_entry, core_guidelines_index } from "../../indexes/core_guidelines/types.js";
-import { BotComponent } from "../bot-component.js";
-import { CommandSetBuilder } from "../command-abstractions/command-set-builder.js";
-import { Wheatley } from "../wheatley.js";
-import { colors } from "../common.js";
-import { EarlyReplyMode, TextBasedCommandBuilder } from "../command-abstractions/text-based-command-builder.js";
-import { TextBasedCommand } from "../command-abstractions/text-based-command.js";
+import { Index, IndexEntry } from "../../../algorithm/search.js";
+import { core_guidelines_entry, core_guidelines_index } from "../../../../indexes/core_guidelines/types.js";
+import { BotComponent } from "../../../bot-component.js";
+import { CommandSetBuilder } from "../../../command-abstractions/command-set-builder.js";
+import { Wheatley } from "../../../wheatley.js";
+import { colors } from "../../../common.js";
+import { EarlyReplyMode, TextBasedCommandBuilder } from "../../../command-abstractions/text-based-command-builder.js";
+import { TextBasedCommand } from "../../../command-abstractions/text-based-command.js";
 
 type augmented_core_guidelines_entry = core_guidelines_entry & IndexEntry;
