+++ src/modules/tccpp/components/days-since-last-incident.ts
@@ -2,13 +2,13 @@ import * as Discord from "discord.js";
 
 import { strict as assert } from "assert";
 
-import { BotComponent } from "../bot-component.js";
-import { Wheatley } from "../wheatley.js";
-import { unwrap } from "../utils/misc.js";
-import { build_description, capitalize, time_to_human } from "../utils/strings.js";
-import { MINUTE } from "../common.js";
-import { moderation_entry } from "../infra/schemata/moderation.js";
-import { set_interval } from "../utils/node.js";
+import { BotComponent } from "../../../bot-component.js";
+import { Wheatley } from "../../../wheatley.js";
+import { unwrap } from "../../../utils/misc.js";
+import { build_description, capitalize, time_to_human } from "../../../utils/strings.js";
+import { MINUTE } from "../../../common.js";
+import { moderation_entry } from "../../../infra/schemata/moderation.js";
+import { set_interval } from "../../../utils/node.js";
 
 type incident_info = { time: string; user: string; user_name: string; type: string };
