+++ src/modules/tccpp/components/forum-channels.ts
@@ -1,13 +1,13 @@
 import * as Discord from "discord.js";
 import { strict as assert } from "assert";
-import { decode_snowflake, fetch_all_threads_archive_count, get_tag } from "../utils/discord.js";
-import { SelfClearingSet } from "../utils/containers.js";
-import { M } from "../utils/debugging-and-logging.js";
-import { colors, DAY, HOUR, MINUTE } from "../common.js";
-import { BotComponent } from "../bot-component.js";
-import { Wheatley } from "../wheatley.js";
-import { clear_timeout, set_interval, set_timeout } from "../utils/node.js";
-import { Synopsinator } from "../utils/synopsis.js";
+import { decode_snowflake, fetch_all_threads_archive_count, get_tag } from "../../../utils/discord.js";
+import { SelfClearingSet } from "../../../utils/containers.js";
+import { M } from "../../../utils/debugging-and-logging.js";
+import { colors, DAY, HOUR, MINUTE } from "../../../common.js";
+import { BotComponent } from "../../../bot-component.js";
+import { Wheatley } from "../../../wheatley.js";
+import { clear_timeout, set_interval, set_timeout } from "../../../utils/node.js";
+import { Synopsinator } from "../../../utils/synopsis.js";
 
 // TODO: Take into account thread's inactivity setting
