+++ src/modules/tccpp/components/roulette.ts
@@ -1,14 +1,14 @@
 import * as Discord from "discord.js";
 import { strict as assert } from "assert";
-import { SelfClearingMap, SelfClearingSet } from "../utils/containers.js";
-import { M } from "../utils/debugging-and-logging.js";
-import { colors, MINUTE } from "../common.js";
-import { BotComponent } from "../bot-component.js";
-import { CommandSetBuilder } from "../command-abstractions/command-set-builder.js";
-import { Wheatley } from "../wheatley.js";
-import { EarlyReplyMode, TextBasedCommandBuilder } from "../command-abstractions/text-based-command-builder.js";
-import { TextBasedCommand } from "../command-abstractions/text-based-command.js";
-import { roulette_leaderboard_entry } from "../infra/schemata/roulette.js";
+import { SelfClearingMap, SelfClearingSet } from "../../../utils/containers.js";
+import { M } from "../../../utils/debugging-and-logging.js";
+import { colors, MINUTE } from "../../../common.js";
+import { BotComponent } from "../../../bot-component.js";
+import { CommandSetBuilder } from "../../../command-abstractions/command-set-builder.js";
+import { Wheatley } from "../../../wheatley.js";
+import { EarlyReplyMode, TextBasedCommandBuilder } from "../../../command-abstractions/text-based-command-builder.js";
+import { TextBasedCommand } from "../../../command-abstractions/text-based-command.js";
+import { roulette_leaderboard_entry } from "../../../infra/schemata/roulette.js";
 
 const LEADERBOARD_ENTRIES = 20;
