+++ src/modules/tccpp/components/status.ts
@@ -2,8 +2,8 @@ import * as Discord from "discord.js";
 
 import { strict as assert } from "assert";
 
-import { BotComponent } from "../bot-component.js";
-import { Wheatley } from "../wheatley.js";
+import { BotComponent } from "../../../bot-component.js";
+import { Wheatley } from "../../../wheatley.js";
 
 export default class Status extends BotComponent {
     override async on_ready() {
