+++ src/modules/tccpp/components/thread-based-channels.ts
@@ -1,10 +1,10 @@
 import * as Discord from "discord.js";
 import { strict as assert } from "assert";
-import { unwrap } from "../utils/misc.js";
-import { M } from "../utils/debugging-and-logging.js";
-import { colors } from "../common.js";
-import { BotComponent } from "../bot-component.js";
-import { Wheatley } from "../wheatley.js";
+import { unwrap } from "../../../utils/misc.js";
+import { M } from "../../../utils/debugging-and-logging.js";
+import { colors } from "../../../common.js";
+import { BotComponent } from "../../../bot-component.js";
+import { Wheatley } from "../../../wheatley.js";
 
 /*
  * Thread-based channel logic (non-forum)
